chore: seperating out trained model tests
