+++ tests/test_pipeline.py
@@ -1,19 +1,16 @@
 """Tests for Pipeline."""
 
-import numpy as np
 import pytest
 
-from sotai import FeatureType, Metric, Pipeline, TargetType, TrainedModel
-from sotai.models import CalibratedLinear
+from sotai import FeatureType, Metric, Pipeline, TargetType
 
-from .fixtures import (
+from .fixtures import (  # pylint: disable=unused-import
     fixture_test_categories,
     fixture_test_data,
     fixture_test_feature_configs,
     fixture_test_feature_names,
     fixture_test_target,
 )
-from .utils import construct_trained_model
 
 
 @pytest.mark.parametrize(
@@ -149,47 +146,3 @@ def test_pipeline_save_load(
         assert loaded_dataset.prepared_data.train.equals(dataset.prepared_data.train)
         assert loaded_dataset.prepared_data.val.equals(dataset.prepared_data.val)
         assert loaded_dataset.prepared_data.test.equals(dataset.prepared_data.test)
-
-
-def test_trained_classification_model_predict(
-    test_data: fixture_test_data, test_feature_configs: fixture_test_feature_configs
-):
-    """Tests the predict function on a trained model."""
-    trained_model = construct_trained_model(
-        TargetType.CLASSIFICATION, test_data, test_feature_configs
-    )
-    predictions, probabilities = trained_model.predict(test_data)
-    assert isinstance(predictions, np.ndarray)
-    assert len(predictions) == len(test_data)
-    assert isinstance(probabilities, np.ndarray)
-    assert len(probabilities) == len(test_data)
-
-
-def test_trained_regression_model_predict(
-    test_data: fixture_test_data, test_feature_configs: fixture_test_feature_configs
-):
-    """Tests the predict function on a trained model."""
-    trained_model = construct_trained_model(
-        TargetType.REGRESSION, test_data, test_feature_configs
-    )
-    predictions = trained_model.predict(test_data)
-    assert isinstance(predictions, np.ndarray)
-    assert len(predictions) == len(test_data)
-
-
-def test_trained_model_save_load(
-    test_data: fixture_test_data,
-    test_feature_configs: fixture_test_feature_configs,
-    tmp_path,
-):
-    """Tests that a `TrainedModel` can be successfully saved and then loaded."""
-    trained_model = construct_trained_model(
-        TargetType.CLASSIFICATION, test_data, test_feature_configs
-    )
-    trained_model.save(tmp_path)
-    loaded_trained_model = TrainedModel.load(tmp_path)
-    assert isinstance(loaded_trained_model, TrainedModel)
-    assert loaded_trained_model.dict(exclude={"model"}) == trained_model.dict(
-        exclude={"model"}
-    )
-    assert isinstance(loaded_trained_model.model, CalibratedLinear)
